feat: :sparkles: Clase Menus
Se creo la clase menus para organizar los diferentes opciones del programa en métodos.

+++ src/main/java/org/laLiga/controlador/Informes.java
@@ -43,10 +43,10 @@ public class Informes {
         return total;
     }
 
-    public int promedio(List<Equipo> arrayEquipos){
+    public float promedio(List<Equipo> arrayEquipos){
         int totalEquipos = arrayEquipos.size();
         int totalGoles = this.totalGoles(arrayEquipos);
-        int total = totalGoles/totalEquipos;
+        float total = totalGoles/totalEquipos;
         return total;
     }
 }
